Added coarse-grained generic diffing.
Generic diffs currently only check existence. That is, if the
exact generic type isn't found in another struct's declaration,
it's counted as added or removed.

The next step is to include modifications to a type (like T: Debug +
Clone), and to be able to say "T was modified, added Clone constraint").