Added coarse-grained where clause diffing.
'where' clauses are diffed by existence (for now), similar to generic
types.

Also like generics, the plan is to introduce per-type/lifetime diffs on
the constraints, but that's a future problem. The immediate goal is to
get an end-to-end diff of all main components (struct args, return
types, etc), before tackling more specific diff problems.